third_party/blink/web_tests/fast/css/image-set-parsing-expected.txt

Test the parsing of the image-set functions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Single value for background-image : image-set(url("#a") 1x)
PASS result is "image-set(url(\"#a\") 1x)"


Single value for background-image : -webkit-image-set(url("#a") 1x)
PASS result is "image-set(url(\"#a\") 1x)"


Multiple values for background-image : image-set(url("#a") 1x, url("#b") 2x)
PASS result is "image-set(url(\"#a\") 1x, url(\"#b\") 2x)"


Multiple values for background-image : -webkit-image-set(url("#a") 1x, url("#b") 2x)
PASS result is "image-set(url(\"#a\") 1x, url(\"#b\") 2x)"


Multiple values for background-image, out of order : image-set(url("#c") 3x, url("#b") 2x, url("#a") 1x)
PASS result is "image-set(url(\"#c\") 3x, url(\"#b\") 2x, url(\"#a\") 1x)"


Multiple values for background-image, out of order : -webkit-image-set(url("#c") 3x, url("#b") 2x, url("#a") 1x)
PASS result is "image-set(url(\"#c\") 3x, url(\"#b\") 2x, url(\"#a\") 1x)"


Single value for content : image-set(url("#a") 1x)
PASS result is "image-set(url(\"#a\") 1x)"


Single value for content : -webkit-image-set(url("#a") 1x)
PASS result is "image-set(url(\"#a\") 1x)"


Multiple values for content : image-set(url("#a") 1x, url("#b") 2x)
PASS result is "image-set(url(\"#a\") 1x, url(\"#b\") 2x)"


Multiple values for content : -webkit-image-set(url("#a") 1x, url("#b") 2x)
PASS result is "image-set(url(\"#a\") 1x, url(\"#b\") 2x)"


Single value for border-image : image-set(url("#a") 1x)
PASS result is "image-set(url(\"#a\") 1x)"


Single value for border-image : -webkit-image-set(url("#a") 1x)
PASS result is "image-set(url(\"#a\") 1x)"


Multiple values for border-image : image-set(url("#a") 1x, url("#b") 2x)
PASS result is "image-set(url(\"#a\") 1x, url(\"#b\") 2x)"


Multiple values for border-image : -webkit-image-set(url("#a") 1x, url("#b") 2x)
PASS result is "image-set(url(\"#a\") 1x, url(\"#b\") 2x)"


Single value for -webkit-mask-box-image-source : image-set(url("#a") 1x)
PASS result is "image-set(url(\"#a\") 1x)"


Single value for -webkit-mask-box-image-source : -webkit-image-set(url("#a") 1x)
PASS result is "image-set(url(\"#a\") 1x)"


Multiple values for -webkit-mask-box-image-source : image-set(url("#a") 1x, url("#b") 2x)
PASS result is "image-set(url(\"#a\") 1x, url(\"#b\") 2x)"


Multiple values for -webkit-mask-box-image-source : -webkit-image-set(url("#a") 1x, url("#b") 2x)
PASS result is "image-set(url(\"#a\") 1x, url(\"#b\") 2x)"
PASS successfullyParsed is true

TEST COMPLETE
